fix paswword validator

+++ Back/controllers/mainpage-controllers.js
@@ -66,7 +66,9 @@ exports.updatePost = (req, res) => {
         Post.findOne({ _id: req.params.id })
             .then((PostArg) => {
                 const filename = PostArg.picture.split('/images/uploads/posts/')[1];
-                if (filename !== 'undefined') {
+                console.log('test')
+                console.log(PostArg.picture === '')
+                if (filename !== 'undefined' && PostArg.picture !== '') {
                     fs.unlink(`images/uploads/posts/${filename}`, (error) => {
                         if (error) throw error;
                     });
